requirements.txt
fastapi==0.104.1
uvicorn==0.23124589*/.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart>=0.0.7
numpy>=2.2.6
python-dotenv>=1.0.0